refactor(*): switch to a monorepo
