--- esm/index.mjs
@@ -1,4 +1,0 @@
-import Endb from '../src/index.js';
-
-export default Endb;
-export { Endb };
