+++ jest.config.js
@@ -1,8 +1,8 @@
 module.exports = {
   collectCoverage: true,
-  coveragePathIgnorePatterns: ['/node_modules/', '/test/'],
   coverageDirectory: "coverage",
   testEnvironment: "node",
-  testMatch: ["**/test/**/*.test.js"],
+  roots: ['<rootDir>packages'],
+  testMatch: ['**/test/**/*.js'],
   verbose: true,
 };
