+++ packages/endb-mongo/README.md
@@ -1,0 +1,11 @@
+# `endb-mongo`
+
+> TODO: description
+
+## Usage
+
+```
+const endbMongo = require('endb-mongo');
+
+// TODO: DEMONSTRATE API
+```
