+++ packages/endb-mongo/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "@endb/mongo",
+  "version": "1.0.0",
+  "description": "MongoDB adapter for Endb",
+  "main": "src/index.js",
+  "files": [
+    "src"
+  ],
+  "keywords": [
+    "endb",
+    "mongodb"
+  ],
+  "license": "MIT",
+  "author": "chroventer (https://github.com/chroventer)",
+  "homepage": "https://endb.js.org",
+  "repository": "github:endbjs/endb",
+  "bugs": "https://github.com/endbjs/endb/issues",
+  "dependencies": {
+    "mongodb": "^3.5.8"
+  },
+  "devDependencies": {
+    "endb": "*",
+    "@endb/test": "*"
+  }
+}
