+++ packages/endb-mysql/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "@endb/mysql",
+  "version": "1.0.0",
+  "description": "MySQL adapter for Endb",
+  "main": "src/index.js",
+  "files": ["src"],
+  "license": "MIT",
+  "author": "chroventer (https://github.com/chroventer)",
+  "homepage": "https://endb.js.org",
+  "repository": "github:endbjs/endb",
+  "bugs": "https://github.com/endbjs/endb/issues",
+  "dependencies": {
+    "mysql2": "2.1.0",
+    "@endb/sql": "*"
+  },
+  "devDependencies": {
+    "endb": "*",
+    "@endb/test": "*"
+  }
+}
