+++ packages/endb-postgres/README.md
@@ -1,0 +1,11 @@
+# `endb-postgres`
+
+> TODO: description
+
+## Usage
+
+```
+const endbPostgres = require('endb-postgres');
+
+// TODO: DEMONSTRATE API
+```
