+++ packages/endb-redis/README.md
@@ -1,0 +1,11 @@
+# `endb-redis`
+
+> TODO: description
+
+## Usage
+
+```
+const endbRedis = require('endb-redis');
+
+// TODO: DEMONSTRATE API
+```
