+++ packages/endb-redis/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "@endb/redis",
+  "version": "1.0.0",
+  "description": "Redis adapter for Endb",
+  "main": "src/index.js",
+  "files": ["src"],
+  "license": "MIT",
+  "author": "chroventer (https://github.com/chroventer)",
+  "homepage": "https://endb.js.org",
+  "repository": "github:endbjs/endb",
+  "bugs": "https://github.com/endbjs/endb/issues",
+  "dependencies": {
+    "redis": "3.0.2"
+  },
+  "devDependencies": {
+    "endb": "*",
+    "@endb/test": "*"
+  }
+}
