+++ packages/endb-redis/test/index.js
@@ -1,0 +1,12 @@
+'use strict';
+
+const Endb = require('endb');
+const EndbRedis = require('..');
+const { endbTest, adapterTest } = require('@endb/test');
+const { REDIS_HOST = 'localhost' } = process.env;
+const uri = `redis://${REDIS_HOST}`;
+
+adapterTest(test, Endb, uri, 'redis://foo');
+endbTest(test, Endb, {
+    store: new EndbRedis(uri)
+});
