+++ packages/endb-sql/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "@endb/sql",
+  "version": "1.0.0",
+  "description": "SQL adapter for Endb",
+  "main": "src/index.js",
+  "files": ["src"],
+  "license": "MIT",
+  "author": "chroventer (https://github.com/chroventer)",
+  "homepage": "https://endb.js.org",
+  "repository": "github:endbjs/endb",
+  "bugs": "https://github.com/endbjs/endb/issues",
+  "dependencies": {
+    "sql": "0.78.0"
+  },
+  "devDependencies": {
+    "endb": "*",
+    "@endb/test": "*",
+    "sqlite3": "4.2.0"
+  }
+}
