+++ packages/endb-test/README.md
@@ -1,0 +1,11 @@
+# `endb-test`
+
+> TODO: description
+
+## Usage
+
+```
+const endbTest = require('endb-test');
+
+// TODO: DEMONSTRATE API
+```
