+++ packages/endb/README.md
@@ -1,0 +1,11 @@
+# `endb`
+
+> TODO: description
+
+## Usage
+
+```
+const endb = require('endb');
+
+// TODO: DEMONSTRATE API
+```
