--- test/adapters/mongodb.test.js
@@ -1,9 +1,0 @@
-'use strict';
-
-const Endb = require('../../src');
-const { endbTest, adapterTest } = require('../functions');
-const { MONGO_HOST = '127.0.0.1' } = process.env;
-const uri = `mongodb://${MONGO_HOST}:27017`;
-
-adapterTest(test, Endb, uri, 'mongodb://127.0.0.1:1234');
-endbTest(test, Endb, { uri });
